Refactoring of Aspect
